+++ src/components/index.js
@@ -1,1 +1,3 @@
+export { EditTestsFilter } from './EditTestsFilter/EditTestsFilter';
+export { EditTestsTable } from './EditTestsTable/EditTestsTable';
 export { TestLevelSelector } from './TestLevelSelector/TestLevelSelector';
