+++ src/pages/EditTests/EditTests.scss
@@ -1,5 +1,5 @@
 .edit-tests-selector {
-  margin: 60px 0;
+  margin: 60px auto;
 
   .edit-tests-placeholder {
     display: flex;
